requirements.txt
requests==2.31.0
python-dotenv==1.0.0
xmltodict==0.13.0
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
Flask==2.3.3
Werkzeug==2.3.7
SQLAlchemy==2.0.23
google-generativeai==0.3.1
gunicorn==21.2.0 